+++ src/index.ts
@@ -37,7 +37,7 @@ export {
 
 export { JsonTestLogger } from './logger/JsonTestLogger'
 export { NoConsole } from './logger/NoConsole'
-export { NoLogger, NoOpTestLogger } from './logger/NoLogger'
+export { NoOpLogger, NoOpTestLogger } from './logger/NoOpLogger'
 export { TextTestLogger } from './logger/TextTestLogger'
 
 export { generateGetParametersFromGraphQLRequestInfo } from './request/GenerateGetParametersFromGraphQLRequestInfo'
